chore(general): fixed load_experiment.py
to pass test cases by also excluding keys that are scan names which were being included in measured. Ultimately we may want to fix this so that we have scans as sub items of the runinfo key within our loaded dictionary. This also could cause issues for measured items that have the same names as these blacklisted scans, such as 'repeat', or 'v1_voltage', and will not exclude scans of other names from loading in measured (not changed in this commit but important to know before merging).